requirements.txt
streamlit
mermaid-python
pandas
numpy
watchdog
PyYAML
